Remove viewing report full screen, only allow printing
