+++ front/app/containers/Admin/reporting/components/ReportBuilderPage/ReportRow/index.tsx
@@ -44,10 +44,6 @@ const ReportRow = ({ report }: Props) => {
     clHistory.push(`${reportPath}/editor`);
   };
 
-  const handleViewReport = () => {
-    window.open(`${reportPath}/viewer`, '_blank');
-  };
-
   return (
     <ListItem>
       <Box
@@ -73,7 +69,6 @@ const ReportRow = ({ report }: Props) => {
           canEdit={report.attributes.action_descriptor.editing_report.enabled}
           onDelete={handleDeleteReport}
           onEdit={handleEditReport}
-          onView={handleViewReport}
         />
       </Box>
     </ListItem>
