+++ front/app/routes.tsx
@@ -43,9 +43,6 @@ const SubscriptionEndedPage = lazy(
 );
 const EmailSettingsPage = lazy(() => import('containers/EmailSettingsPage'));
 
-const ReportViewerPage = lazy(
-  () => import('containers/Admin/reporting/containers/FullScreenReport')
-);
 const ReportPrintPage = lazy(
   () => import('containers/Admin/reporting/containers/PrintReport')
 );
@@ -323,14 +320,6 @@ export default function createRoutes() {
           element: (
             <PageLoading>
               <EmailSettingsPage />
-            </PageLoading>
-          ),
-        },
-        {
-          path: 'admin/reporting/report-builder/:reportId/viewer',
-          element: (
-            <PageLoading>
-              <ReportViewerPage />
             </PageLoading>
           ),
         },
